Adds .csproj for Windows files to simplify linking process
Linking the react-native-share module as a project instead of as source will make for a simpler process.  Now, the user need only run `npm install --save react-native-share`, add the .csproj in the node_modules folder to their solution, reference from their app, and add the necessary code to MainPage.cs.

I've also added RNShare.sln to confirm that the files (RNShareModule.cs and RNSharePackage.cs) can compile.  To use this, first install `react-native-windows` at the root of the project, then open RNShare.sln and build using the Development configuration.  The Development configuration configures the location of the ReactNative.csproj differently from Debug and Release configurations, which are configured as if `react-native-windows` is a peer dependency as opposed to a direct dependency.